mark pointer_from_buffer as unsafe (#9)
For safety, pointer_from_buffer must never be called more than once for a given call to pointer_to_buffer. To prevent leaks, it must only be called once.

This makes the function internal as it is not intended to be exposed.
It marks it unsafe so callers need to use unsafe to call it.
A drop implementation is added to IdPointerMessage in case into_id_box is never called, to prevent leaks.